% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\docType{package}
\name{forecastexp-package}
\alias{forecastexp-package}
\title{forecastexp: General purpose forecast experiment framework}
\description{
forecastexp: General purpose forecast experiment framework
}

